pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "autotest-aiops-agent"
version = "0.1.0"
description = "AutoTest AIOps Agent"
dependencies = [
    "langchain",
    "langchain-openai",
    "langgraph",
    "llama-index",
    "llama-index-vector-stores-chroma",
    "llama-index-embeddings-huggingface",
    "chromadb",
    "drain3",
    "docling",
    "watchdog",
    "openai",
    "sentence-transformers",
    "slack-sdk",
    "psutil",
    "python-dotenv"
]
requires-python = ">=3.8"

[tool.setuptools.packages.find]
include = ["agent*"]